Restructure plugin.
