--- src/blocks/plugin-stats/constants.js
@@ -1,8 +1,0 @@
-// Allowed formats for the prefix and suffix fields.
-export const ALLOWED_FORMATS = [
-	'core/bold',
-	'core/italic',
-	'core/link',
-	'core/strikethrough',
-	'core/text-color',
-];
